refactor: relocate Next impls to their defining type modules
Move the repeat-last Next inherent methods out of the feature-named
next.rs orphan files into the modules that define each type:

- Point::eval_next            -> point.rs (beside eval_eq/eval_select)
- Poly::{eval_next_base,new_next_from_point} -> poly.rs
- EqMaybePacked next helpers  -> split_eq/eq.rs
- SplitEq next views          -> split_eq/mod.rs
- SvoPoint / NextPartials next logic + helpers -> svo.rs

This also fixes the visibility leak where pub(super) helpers written in
crate::next resolved to the whole crate; under split_eq they now scope
to the split_eq module as intended. Adds #[inline] to Point::eval_next
for parity with its siblings. Pure relocation; tests move alongside,
no behavior change.
