+++ multilinear-util/src/lib.rs
@@ -3,7 +3,6 @@
 extern crate alloc;
 
 pub mod eq_batch;
-pub mod next;
 pub mod point;
 pub mod poly;
 pub mod split_eq;
